Russian APT/RomCom_Storm0978/requirements.txt
pycryptodome>=3.19.0
requests>=2.31.0
psutil>=5.9.0
